requirements.txt
kafka-python==1.4.2
boto3
